requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
PyJWT==2.8.0
redis==5.0.1
supabase==2.0.4
openai==1.3.5
python-dotenv==1.0.0
requests==2.31.0
sentry-sdk[flask]==1.38.0
mixpanel==4.10.0
gunicorn==21.2.0
psycopg2-binary==2.9.7